Prevent floating point exception from ADDING uint64 and double
Observed: EInvalidOp exception from adding very large (close to max)
uint64_t value and double value = 40000.
Really, just from adding, there was no division by zero in tests.

+++ baresip/modules/colorbar_generator/colorbar_generator.c
@@ -30,7 +30,7 @@ static struct vidsrc *vidsrc;
 
 static void process_frame(struct vidsrc_st *st)
 {
-	st->ts += (VIDEO_TIMEBASE / st->fps);
+	st->ts += (uint64_t)(VIDEO_TIMEBASE / st->fps);
 
 	st->frameh(st->frame, /*st->ts, */ st->arg);
 }
